InitializeUninitialized: fix initializing array allocas
we do not want to store anything there, since the size may be 0.
Instead of storing stuff there, just call klee_make_symbolic.
We do not do this for simple types, because we want the slicer
to delete those stores when the memory is initialized later
(so that KLEE may have less symbolic objects)