do not copy and transform sources
If the sources include some local files, that if we copy them,
we need to copy also the included files and we do not want to
do that. Instead, just compile those files directly and save
the output to the working directory.
This also means that we can not do source transformations,
but we wnated to get rid of them anyway.