utils/process.py: Fix possible discarding of subprocess output
ProcessRunner may discard some parts of a sub-process output whenever
the sub-process writes of the pipe and subsequently exits in the exact
moment between the lines below.

```
line = ProcessRunner.current_process.stdout.readline()
if line == b'' and ProcessRunner.current_process.poll() is not None:
```

This behaviour could be observed when multiple instances of Symbiotic
were launched in parallel.

Fixes #138.