Use -break-infinite-loops before slicing for termination
We commented as non-needed, but termination needs it - it creates an
artificial exit that is then marked by -find-exits and we can slice
w.r.t to it. Without the breaking, we do not have a point to
which we can slice if the loop is while(true) as LLVM keeps only
the reachable blocks.